+++ apps/public-docsite-v9/.storybook/main.js
@@ -1,5 +1,16 @@
+// @ts-check
+
+const {
+  getPackageStoriesGlob,
+  createPathAliasesConfig,
+  registerTsPaths,
+  rules,
+  registerRules,
+} = require('@fluentui/scripts-storybook');
+
 const rootMain = require('../../../.storybook/main');
-const { getPackageStoriesGlob } = require('@fluentui/scripts-storybook');
+
+const { tsConfigAllPath } = createPathAliasesConfig();
 
 module.exports = /** @type {Omit<import('../../../.storybook/main'), 'typescript'|'babel'>} */ ({
   ...rootMain,
@@ -14,9 +25,11 @@ module.exports = /** @type {Omit<import('../../../.storybook/main'), 'typescript
   staticDirs: ['../public'],
   addons: [...rootMain.addons],
   webpackFinal: (config, options) => {
-    const localConfig = { ...rootMain.webpackFinal(config, options) };
+    const localConfig = /** @type config */ ({ ...rootMain.webpackFinal(config, options) });
 
     // add your own webpack tweaks if needed
+    registerTsPaths({ configFile: tsConfigAllPath, config: localConfig });
+    registerRules({ rules: [rules.scssRule], config: localConfig });
 
     return localConfig;
   },