+++ packages/react-components/react-table/src/index.ts
@@ -1,4 +1,11 @@
-export { useTableFeatures, useTableSelection, useTableSort, createTableColumn } from './hooks';
+export {
+  useTableFeatures,
+  useTableSelection,
+  useTableSort,
+  createTableColumn,
+  useTableColumnSizing_unstable,
+} from './hooks';
+
 export type {
   CreateTableColumnOptions,
   UseTableFeaturesOptions,
@@ -10,6 +17,7 @@ export type {
   TableSelectionState,
   TableSortState,
   TableFeaturePlugin,
+  TableColumnSizingOptions,
 } from './hooks';
 
 export {
@@ -71,6 +79,15 @@ export {
   renderTableHeaderCell_unstable,
 } from './TableHeaderCell';
 export type { TableHeaderCellProps, TableHeaderCellSlots, TableHeaderCellState } from './TableHeaderCell';
+
+export {
+  TableResizeHandle,
+  tableResizeHandleClassNames,
+  useTableResizeHandleStyles_unstable,
+  useTableResizeHandle_unstable,
+  renderTableResizeHandle_unstable,
+} from './TableResizeHandle';
+export type { TableResizeHandleProps, TableResizeHandleSlots, TableResizeHandleState } from './TableResizeHandle';
 
 export { TableContextProvider, useTableContext } from './contexts/tableContext';
 export { useTableRowIdContext, TableRowIdContextProvider } from './contexts/rowIdContext';