+++ scripts/storybook/src/index.d.ts
@@ -1,1 +1,10 @@
-export { getPackageStoriesGlob, getCodesandboxBabelOptions, loadWorkspaceAddon, registerTsPaths } from './utils';
+export {
+  getPackageStoriesGlob,
+  loadWorkspaceAddon,
+  registerTsPaths,
+  registerRules,
+  createPathAliasesConfig,
+  overrideDefaultBabelLoader,
+} from './utils';
+
+export * as rules from './rules';