+++ scripts/storybook/src/utils.js
@@ -2,9 +2,9 @@ const fs = require('fs');
 const path = require('path');
 
 const { isConvergedPackage, getAllPackageInfo, getProjectMetadata } = require('@fluentui/scripts-monorepo');
-const { stripIndents, offsetFromRoot } = require('@nrwl/devkit');
-const { workspaceRoot } = require('nx/src/utils/app-root');
+const { stripIndents, offsetFromRoot, workspaceRoot, readJsonFile, writeJsonFile } = require('@nrwl/devkit');
 const semver = require('semver');
+const { babelPlugin } = require('storybook-addon-export-to-codesandbox');
 const { TsconfigPathsPlugin } = require('tsconfig-paths-webpack-plugin');
 
 const loadWorkspaceAddonDefaultOptions = { workspaceRoot };
@@ -133,22 +133,44 @@ function loadWorkspaceAddon(addonName, options) {
 }
 
 /**
- * @returns {import('storybook-addon-export-to-codesandbox').BabelPluginOptions}
+ * @private
+ * @param {ReturnType<typeof getAllPackageInfo>} allPackageInfo
+ * @returns {import("webpack").RuleSetRule}
  */
-function getCodesandboxBabelOptions() {
-  const allPackageInfo = getAllPackageInfo();
+function _createCodesandboxRule(allPackageInfo = getAllPackageInfo()) {
+  return {
+    /**
+     * why the usage of 'post' ? - we need to run this loader after all storybook webpack rules/loaders have been executed.
+     * while we can use Array.prototype.unshift to "override" the indexes this approach is more declarative without additional hacks.
+     */
+    enforce: 'post',
+    test: /\.stories\.tsx$/,
+    include: /stories/,
+    exclude: /node_modules/,
+    use: {
+      loader: 'babel-loader',
+      options: _processBabelLoaderOptions({
+        plugins: [[babelPlugin, getCodesandboxBabelOptions()]],
+      }),
+    },
+  };
 
-  return Object.values(allPackageInfo).reduce((acc, cur) => {
-    if (isConvergedPackage({ packagePathOrJson: cur.packageJson, projectType: 'library' })) {
-      const isPrerelease = semver.prerelease(cur.packageJson.version) !== null;
+  /**
+   * @returns {import('storybook-addon-export-to-codesandbox').BabelPluginOptions}
+   */
+  function getCodesandboxBabelOptions() {
+    return Object.values(allPackageInfo).reduce((acc, cur) => {
+      if (isConvergedPackage({ packagePathOrJson: cur.packageJson, projectType: 'library' })) {
+        const isPrerelease = semver.prerelease(cur.packageJson.version) !== null;
 
-      acc[cur.packageJson.name] = isPrerelease
-        ? { replace: '@fluentui/react-components/unstable' }
-        : { replace: '@fluentui/react-components' };
-    }
+        acc[cur.packageJson.name] = isPrerelease
+          ? { replace: '@fluentui/react-components/unstable' }
+          : { replace: '@fluentui/react-components' };
+      }
 
-    return acc;
-  }, /** @type import('storybook-addon-export-to-codesandbox').BabelPluginOptions*/ ({}));
+      return acc;
+    }, /** @type import('storybook-addon-export-to-codesandbox').BabelPluginOptions*/ ({}));
+  }
 }
 
 /**
@@ -190,23 +212,164 @@ function getPackageStoriesGlob(options) {
  *
  * register TsconfigPathsPlugin to webpack config
  * @param {Object} options
- * @param {string} options.tsConfigPath - absolute path to tsconfig that contains path aliases
- * @param {import('webpack').Configuration} options.config
+ * @param {string} options.configFile - absolute path to tsconfig that contains path aliases
+ * @param {import('webpack').Configuration} options.config - webpack config
  * @returns
  */
 function registerTsPaths(options) {
-  const { config, tsConfigPath } = options;
+  const { config, configFile } = options;
   const tsPaths = new TsconfigPathsPlugin({
-    configFile: tsConfigPath,
+    configFile,
   });
 
   config.resolve = config.resolve ?? {};
   config.resolve.plugins = config.resolve.plugins ?? [];
+
+  // remove existing to prevent multiple tspaths plugin
+  config.resolve.plugins = config.resolve.plugins.filter(plugin => !(plugin instanceof TsconfigPathsPlugin));
+
   config.resolve.plugins.push(tsPaths);
+
   return config;
+}
+
+/**
+ *
+ * register custom Webpack Rules to webpack config
+ * @param {Object} options
+ * @param {import('webpack').RuleSetRule[]} options.rules - webpack rules
+ * @param {import('webpack').Configuration} options.config - webpack config
+ * @returns
+ */
+function registerRules(options) {
+  const { config, rules } = options;
+  config.module = config.module ?? {};
+  config.module.rules = config.module.rules ?? [];
+  config.module.rules.push(...rules);
+
+  return config;
+}
+
+/**
+ * @typedef {import('@babel/core').TransformOptions & Partial<{customize: string | null}>} BabelLoaderOptions
+ */
+
+/**
+ * Adds custom config to any `babel-loader` usage. Needs to be used on all manually added rules with babel-loader to webpack configuration.
+ *
+ * Why is this needed:
+ *  - `options.babelrc` is ignored by `babel-loader` thus we need to use `customize` api to exclude specific babel presets/plugins
+ *
+ * @private
+ * @param {BabelLoaderOptions} loaderConfig
+ */
+function _processBabelLoaderOptions(loaderConfig) {
+  const customLoaderPath = path.join(__dirname, './loaders/custom-loader.js');
+  const customOptions = { customize: customLoaderPath };
+  Object.assign(loaderConfig, customOptions);
+
+  return loaderConfig;
+}
+
+/**
+ * @typedef  {{loader: string; options: { [index: string]: any }}} LoaderObjectDef
+ */
+
+/**
+ * Overrides storybooks babel-loader setup
+ *
+ * We might remove this once we'll came up with robust solution (or proper behaviors will be added to babel-loader). For more context @see https://github.com/microsoft/fluentui/issues/18775
+ *
+ * 📣 We don't use this override anymore as babel-loader is replaced by swc in whole webpack via `storybook-addon-swc`
+ *
+ * **Note:**
+ * - this function mutates `rules` argument which is a reference to `modules.rules` webpack config property
+ * - to print used babel-loader config run: `yarn start-storybook --no-manager-cache --debug-webpack` and look for
+ * webpack rule set containing both:
+ *  - `test: /\.(mjs|tsx?|jsx?)$/`
+ *  - `node_modules/babel-loader/lib/index.js` as `loader` within module.rules
+ *
+ * @param {Object} options
+ * @param {import('webpack').Configuration} options.config - webpack config
+ */
+function overrideDefaultBabelLoader(options) {
+  const { config } = options;
+  config.module = config.module ?? {};
+  config.module.rules = config.module.rules ?? [];
+
+  const loader = getBabelLoader(/** @type {import('webpack').RuleSetRule[]}*/ (config.module.rules));
+
+  _processBabelLoaderOptions(loader.options);
+
+  function getBabelLoader(/** @type {import('webpack').RuleSetRule[]} */ rules) {
+    // eslint-disable-next-line no-shadow
+    const ruleIdx = rules.findIndex(rule => {
+      return String(rule.test) === '/\\.(mjs|tsx?|jsx?)$/';
+    });
+
+    const rule = /** @type {import("webpack").RuleSetRule}*/ (rules[ruleIdx]);
+
+    if (!Array.isArray(rule.use)) {
+      throw new Error('storybook webpack rules changed');
+    }
+
+    const loaderIdx = rule.use.findIndex(loaderConfig => {
+      return /** @type {LoaderObjectDef} */ (loaderConfig).loader.includes('babel-loader');
+    });
+
+    // eslint-disable-next-line no-shadow
+    const loader = /** @type {LoaderObjectDef}*/ (rule.use[loaderIdx]);
+
+    if (!Object.prototype.hasOwnProperty.call(loader, 'options')) {
+      throw new Error('storybook webpack #module.rules changed!');
+    }
+
+    return loader;
+  }
+}
+
+/**
+ * Create tsconfig.json with merged "compilerOptions.paths" from v0,v8,v9 tsconfigs.
+ *
+ * Main purpose of this is to be used for build-less DX in webpack in tandem with {@link registerTsPaths}
+ * @returns
+ */
+function createPathAliasesConfig() {
+  const { tsConfigAllPath } = createMergedTsConfig();
+  return { tsConfigAllPath };
+}
+
+function createMergedTsConfig() {
+  const rootPath = workspaceRoot;
+  const tsConfigAllPath = path.join(rootPath, 'dist/tsconfig.base.all.json');
+  const baseConfigs = {
+    v0: readJsonFile(path.join(rootPath, 'tsconfig.base.v0.json')),
+    v8: readJsonFile(path.join(rootPath, 'tsconfig.base.v8.json')),
+    v9: readJsonFile(path.join(rootPath, 'tsconfig.base.json')),
+  };
+  const mergedTsConfig = {
+    compilerOptions: {
+      moduleResolution: 'node',
+      forceConsistentCasingInFileNames: true,
+      skipLibCheck: true,
+      baseUrl: workspaceRoot,
+      paths: {
+        ...baseConfigs.v0.compilerOptions.paths,
+        ...baseConfigs.v8.compilerOptions.paths,
+        ...baseConfigs.v9.compilerOptions.paths,
+      },
+    },
+  };
+
+  writeJsonFile(tsConfigAllPath, mergedTsConfig);
+
+  return { tsConfigAllPath, mergedTsConfig };
 }
 
 exports.getPackageStoriesGlob = getPackageStoriesGlob;
 exports.loadWorkspaceAddon = loadWorkspaceAddon;
-exports.getCodesandboxBabelOptions = getCodesandboxBabelOptions;
 exports.registerTsPaths = registerTsPaths;
+exports.registerRules = registerRules;
+exports.createPathAliasesConfig = createPathAliasesConfig;
+exports.overrideDefaultBabelLoader = overrideDefaultBabelLoader;
+exports._createCodesandboxRule = _createCodesandboxRule;